pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flow-llm-router"
version = "0.1.0"
description = "Local-first LLM gateway with token usage analytics for AI developers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "FlowGate Team" }]
keywords = ["llm", "gateway", "proxy", "ai", "token-usage"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.22",
    "litellm>=1.70",
    "typer>=0.12",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "cryptography>=43.0",
]

[project.optional-dependencies]
skills = ["chromadb>=0.5"]
classifier = ["routellm[serve]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
flow-router = "flowgate.cli:app"

[project.urls]
Homepage = "https://github.com/user/flow-llm-router"
Documentation = "https://github.com/user/flow-llm-router#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/flowgate"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
